Ensure rabbitmq data directories permissions
RabbitMQ data directories already exists in the Rock, but not when
mounted on the pod. In very rare occurence, the charm seems to be
executing before the mount is ready, where it will see the folder
/var/lib/rabbitmq already existing.

Check if the folder already exists, create the mnesia folder if it's
missing, with proper permissions.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>