Always generate an erlang cookie
When the unit is the leader, and there's no erlang cookie in the
databag, generate a new one. When the new data folder is mounted,
there's no erlang cookie file, and we see in rare occasions rabbitmq
considering this as a permissions access error instead of generating a
new one.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>